EmployeeExistException

+++ patika-06-01/src/main/java/dev/patika/patika0601/service/EmployeeService.java
@@ -1,5 +1,6 @@
 package dev.patika.patika0601.service;
 
+import dev.patika.patika0601.exceptions.EmployeeExistException;
 import dev.patika.patika0601.model.Employee;
 import dev.patika.patika0601.repository.EmployeeRepository;
 import lombok.AllArgsConstructor;
@@ -17,13 +18,21 @@ public class EmployeeService {
     }
 
     public void addEmployee(Employee employee) {
-        // TO-DO check if employee exists???
+        boolean isExists = repository.existsByEmail(employee.getEmail());
+        if(isExists){
+            throw new EmployeeExistException("Employee with email: " + employee.getEmail() + " is already exists!");
+        }
         repository.save(employee);
     }
 
     public void deleteEmployee(Long employeeId) {
-        // TO-DO check if employee exists with given ID???
-        repository.deleteById(employeeId);
+        boolean isExists = repository.existsById(employeeId);
+        if(isExists){
+            repository.deleteById(employeeId);
+        }
+        else{
+            throw new EmployeeExistException("Employee with id: " + employeeId + " is not exists!");
+        }
 
     }
 }
